Move parts of RuntimeType to shared partition (#25176)
