+++ src/System.Private.CoreLib/src/System/RuntimeType.CoreCLR.cs
@@ -48,7 +48,7 @@ namespace System
         FullName,
     }
 
-    internal sealed class RuntimeType : TypeInfo, ICloneable
+    internal sealed partial class RuntimeType : TypeInfo, ICloneable
     {
         #region Definitions
 
@@ -3069,17 +3069,6 @@ namespace System
         #endregion
 
         #region Identity
-        public override Module Module => GetRuntimeModule();
-
-        internal RuntimeModule GetRuntimeModule() => RuntimeTypeHandle.GetModule(this);
-
-        public override Assembly Assembly => GetRuntimeAssembly();
-
-        internal RuntimeAssembly GetRuntimeAssembly() => RuntimeTypeHandle.GetAssembly(this);
-
-        public override RuntimeTypeHandle TypeHandle => new RuntimeTypeHandle(this);
-
-        internal sealed override RuntimeTypeHandle GetTypeHandleInternal() => new RuntimeTypeHandle(this);
 
         public sealed override bool IsCollectible
         {
@@ -3088,67 +3077,6 @@ namespace System
                 RuntimeType thisType = this;
                 return RuntimeTypeHandle.IsCollectible(JitHelpers.GetQCallTypeHandleOnStack(ref thisType)) != Interop.BOOL.FALSE;
             }
-        }
-
-        protected override TypeCode GetTypeCodeImpl()
-        {
-            TypeCode typeCode = Cache.TypeCode;
-
-            if (typeCode != TypeCode.Empty)
-                return typeCode;
-
-            CorElementType corElementType = RuntimeTypeHandle.GetCorElementType(this);
-            switch (corElementType)
-            {
-                case CorElementType.ELEMENT_TYPE_BOOLEAN:
-                    typeCode = TypeCode.Boolean; break;
-                case CorElementType.ELEMENT_TYPE_CHAR:
-                    typeCode = TypeCode.Char; break;
-                case CorElementType.ELEMENT_TYPE_I1:
-                    typeCode = TypeCode.SByte; break;
-                case CorElementType.ELEMENT_TYPE_U1:
-                    typeCode = TypeCode.Byte; break;
-                case CorElementType.ELEMENT_TYPE_I2:
-                    typeCode = TypeCode.Int16; break;
-                case CorElementType.ELEMENT_TYPE_U2:
-                    typeCode = TypeCode.UInt16; break;
-                case CorElementType.ELEMENT_TYPE_I4:
-                    typeCode = TypeCode.Int32; break;
-                case CorElementType.ELEMENT_TYPE_U4:
-                    typeCode = TypeCode.UInt32; break;
-                case CorElementType.ELEMENT_TYPE_I8:
-                    typeCode = TypeCode.Int64; break;
-                case CorElementType.ELEMENT_TYPE_U8:
-                    typeCode = TypeCode.UInt64; break;
-                case CorElementType.ELEMENT_TYPE_R4:
-                    typeCode = TypeCode.Single; break;
-                case CorElementType.ELEMENT_TYPE_R8:
-                    typeCode = TypeCode.Double; break;
-                case CorElementType.ELEMENT_TYPE_STRING:
-                    typeCode = TypeCode.String; break;
-                case CorElementType.ELEMENT_TYPE_VALUETYPE:
-                    if (this == Convert.ConvertTypes[(int)TypeCode.Decimal])
-                        typeCode = TypeCode.Decimal;
-                    else if (this == Convert.ConvertTypes[(int)TypeCode.DateTime])
-                        typeCode = TypeCode.DateTime;
-                    else if (IsEnum)
-                        typeCode = GetTypeCode(Enum.GetUnderlyingType(this));
-                    else
-                        typeCode = TypeCode.Object;
-                    break;
-                default:
-                    if (this == Convert.ConvertTypes[(int)TypeCode.DBNull])
-                        typeCode = TypeCode.DBNull;
-                    else if (this == Convert.ConvertTypes[(int)TypeCode.String])
-                        typeCode = TypeCode.String;
-                    else
-                        typeCode = TypeCode.Object;
-                    break;
-            }
-
-            Cache.TypeCode = typeCode;
-
-            return typeCode;
         }
 
         public override MethodBase? DeclaringMethod
@@ -3169,7 +3097,6 @@ namespace System
         #endregion
 
         #region Hierarchy
-        public override bool IsInstanceOfType(object? o) => RuntimeTypeHandle.IsInstanceOfType(this, o);
 
         public override bool IsSubclassOf(Type type)
         {
@@ -3198,53 +3125,6 @@ namespace System
             return false;
         }
 
-        public override bool IsAssignableFrom(TypeInfo? typeInfo)
-        {
-            if (typeInfo == null) return false;
-            return IsAssignableFrom(typeInfo.AsType());
-        }
-
-        public override bool IsAssignableFrom(Type? c)
-        {
-            if (c is null)
-                return false;
-
-            if (ReferenceEquals(c, this))
-                return true;
-
-            // For runtime type, let the VM decide.
-            if (c.UnderlyingSystemType is RuntimeType fromType)
-            {
-                // both this and c (or their underlying system types) are runtime types
-                return RuntimeTypeHandle.CanCastTo(fromType, this);
-            }
-
-            // Special case for TypeBuilder to be backward-compatible.
-            if (c is System.Reflection.Emit.TypeBuilder)
-            {
-                // If c is a subclass of this class, then c can be cast to this type.
-                if (c.IsSubclassOf(this))
-                    return true;
-
-                if (IsInterface)
-                {
-                    return c.ImplementInterface(this);
-                }
-                else if (IsGenericParameter)
-                {
-                    Type[] constraints = GetGenericParameterConstraints();
-                    for (int i = 0; i < constraints.Length; i++)
-                        if (!constraints[i].IsAssignableFrom(c))
-                            return false;
-
-                    return true;
-                }
-            }
-
-            // For anything else we return false.
-            return false;
-        }
-
 #if FEATURE_TYPEEQUIVALENCE
         // Reflexive, symmetric, transitive.
         public override bool IsEquivalentTo(Type? other)
@@ -3264,55 +3144,6 @@ namespace System
         }
 #endif // FEATURE_TYPEEQUIVALENCE
         
-        public override Type? BaseType => GetBaseType();
-
-        private RuntimeType? GetBaseType()
-        {
-            if (IsInterface)
-                return null;
-
-            if (RuntimeTypeHandle.IsGenericVariable(this))
-            {
-                Type[] constraints = GetGenericParameterConstraints();
-
-                RuntimeType baseType = ObjectType;
-
-                for (int i = 0; i < constraints.Length; i++)
-                {
-                    RuntimeType constraint = (RuntimeType)constraints[i];
-
-                    if (constraint.IsInterface)
-                        continue;
-
-                    if (constraint.IsGenericParameter)
-                    {
-                        GenericParameterAttributes special;
-                        special = constraint.GenericParameterAttributes & GenericParameterAttributes.SpecialConstraintMask;
-
-                        if ((special & GenericParameterAttributes.ReferenceTypeConstraint) == 0 &&
-                            (special & GenericParameterAttributes.NotNullableValueTypeConstraint) == 0)
-                            continue;
-                    }
-
-                    baseType = constraint;
-                }
-
-                if (baseType == ObjectType)
-                {
-                    GenericParameterAttributes special;
-                    special = GenericParameterAttributes & GenericParameterAttributes.SpecialConstraintMask;
-                    if ((special & GenericParameterAttributes.NotNullableValueTypeConstraint) != 0)
-                        baseType = ValueType;
-                }
-
-                return baseType;
-            }
-
-            return RuntimeTypeHandle.GetBaseType(this);
-        }
-
-        public override Type UnderlyingSystemType => this;
-
         #endregion
 
         #region Name
@@ -3351,11 +3182,6 @@ namespace System
 
         #region Attributes
 
-        protected override TypeAttributes GetAttributeFlagsImpl()
-        {
-            return RuntimeTypeHandle.GetAttributes(this);
-        }
-
         public override Guid GUID
         {
             get
@@ -3368,18 +3194,6 @@ namespace System
 
         [MethodImplAttribute(MethodImplOptions.InternalCall)]
         private extern void GetGUID(ref Guid result);
-
-        protected override bool IsContextfulImpl() => false;
-
-        protected override bool IsByRefImpl() => RuntimeTypeHandle.IsByRef(this);
-
-        protected override bool IsPrimitiveImpl() => RuntimeTypeHandle.IsPrimitive(this);
-
-        protected override bool IsPointerImpl() => RuntimeTypeHandle.IsPointer(this);
-
-        protected override bool IsCOMObjectImpl() => RuntimeTypeHandle.IsComObject(this, false);
-
-        public sealed override bool IsByRefLike => RuntimeTypeHandle.IsByRefLike(this);
 
 #if FEATURE_COMINTEROP
         internal override bool IsWindowsRuntimeObjectImpl() => IsWindowsRuntimeObjectType(this);
@@ -3397,19 +3211,6 @@ namespace System
 
         internal bool IsDelegate() => GetBaseType() == typeof(MulticastDelegate);
 
-        protected override bool IsValueTypeImpl()
-        {
-            // We need to return true for generic parameters with the ValueType constraint.
-            // So we cannot use the faster RuntimeTypeHandle.IsValueType because it returns 
-            // false for all generic parameters.
-            if (this == typeof(ValueType) || this == typeof(Enum))
-                return false;
-
-            return IsSubclassOf(typeof(ValueType));
-        }
-
-        protected override bool HasElementTypeImpl() => RuntimeTypeHandle.HasElementType(this);
-
         public override GenericParameterAttributes GenericParameterAttributes
         {
             get
@@ -3425,135 +3226,14 @@ namespace System
             }
         }
 
-        public override bool IsSecurityCritical => true;
-
-        public override bool IsSecuritySafeCritical => false;
-
-        public override bool IsSecurityTransparent => false;
-
         #endregion
 
         #region Arrays
 
         public sealed override bool IsSZArray => RuntimeTypeHandle.IsSZArray(this);
 
-        protected override bool IsArrayImpl() => RuntimeTypeHandle.IsArray(this);
-
-        public override int GetArrayRank()
-        {
-            if (!IsArrayImpl())
-                throw new ArgumentException(SR.Argument_HasToBeArrayClass);
-
-            return RuntimeTypeHandle.GetArrayRank(this);
-        }
-
-        public override Type GetElementType() => RuntimeTypeHandle.GetElementType(this);
-
         internal object[] GetEmptyArray() => Cache.GetEmptyArray();
 
-        #endregion
-
-        #region Enums
-
-        public override string[] GetEnumNames()
-        {
-            if (!IsEnum)
-                throw new ArgumentException(SR.Arg_MustBeEnum, "enumType");
-
-            string[] ret = Enum.InternalGetNames(this);
-
-            // Make a copy since we can't hand out the same array since users can modify them
-            return new ReadOnlySpan<string>(ret).ToArray();
-        }
-
-        public override Array GetEnumValues()
-        {
-            if (!IsEnum)
-                throw new ArgumentException(SR.Arg_MustBeEnum, "enumType");
-
-            // Get all of the values
-            ulong[] values = Enum.InternalGetValues(this);
-
-            // Create a generic Array
-            Array ret = Array.CreateInstance(this, values.Length);
-
-            for (int i = 0; i < values.Length; i++)
-            {
-                object val = Enum.ToObject(this, values[i]);
-                ret.SetValue(val, i);
-            }
-
-            return ret;
-        }
-
-        public override Type GetEnumUnderlyingType()
-        {
-            if (!IsEnum)
-                throw new ArgumentException(SR.Arg_MustBeEnum, "enumType");
-
-            return Enum.InternalGetUnderlyingType(this);
-        }
-
-        public override bool IsEnumDefined(object value)
-        {
-            if (value == null)
-                throw new ArgumentNullException(nameof(value));
-
-            // Check if both of them are of the same type
-            RuntimeType valueType = (RuntimeType)value.GetType();
-
-            // If the value is an Enum then we need to extract the underlying value from it
-            if (valueType.IsEnum)
-            {
-                if (!valueType.IsEquivalentTo(this))
-                    throw new ArgumentException(SR.Format(SR.Arg_EnumAndObjectMustBeSameType, valueType, this));
-
-                valueType = (RuntimeType)valueType.GetEnumUnderlyingType();
-            }
-
-            // If a string is passed in
-            if (valueType == StringType)
-            {
-                // Get all of the Fields, calling GetHashEntry directly to avoid copying
-                string[] names = Enum.InternalGetNames(this);
-                if (Array.IndexOf(names, value) >= 0)
-                    return true;
-                else
-                    return false;
-            }
-
-            // If an enum or integer value is passed in
-            if (Type.IsIntegerType(valueType))
-            {
-                RuntimeType underlyingType = Enum.InternalGetUnderlyingType(this);
-                if (underlyingType != valueType)
-                    throw new ArgumentException(SR.Format(SR.Arg_EnumUnderlyingTypeAndObjectMustBeSameType, valueType, underlyingType));
-
-                ulong[] ulValues = Enum.InternalGetValues(this);
-                ulong ulValue = Enum.ToUInt64(value);
-
-                return (Array.BinarySearch(ulValues, ulValue) >= 0);
-            }
-            else
-            {
-                throw new InvalidOperationException(SR.InvalidOperation_UnknownEnumType);
-            }
-        }
-
-        public override string? GetEnumName(object value)
-        {
-            if (value == null)
-                throw new ArgumentNullException(nameof(value));
-
-            Type valueType = value.GetType();
-
-            if (!(valueType.IsEnum || IsIntegerType(valueType)))
-                throw new ArgumentException(SR.Arg_MustBeEnumBaseTypeOrEnum, nameof(value));
-
-            ulong ulValue = Enum.ToUInt64(value);
-
-            return Enum.GetEnumName(this, ulValue);
-        }
         #endregion
 
         #region Generics
@@ -3634,10 +3314,6 @@ namespace System
             return ret;
         }
 
-        public override bool IsGenericTypeDefinition => RuntimeTypeHandle.IsGenericTypeDefinition(this);
-
-        public override bool IsGenericParameter => RuntimeTypeHandle.IsGenericVariable(this);
-
         public override int GenericParameterPosition
         {
             get
@@ -3648,18 +3324,6 @@ namespace System
                 return new RuntimeTypeHandle(this).GetGenericVariableIndex();
             }
         }
-
-        public override Type GetGenericTypeDefinition()
-        {
-            if (!IsGenericType)
-                throw new InvalidOperationException(SR.InvalidOperation_NotGenericType);
-
-            return RuntimeTypeHandle.GetGenericTypeDefinition(this);
-        }
-
-        public override bool IsGenericType => RuntimeTypeHandle.HasInstantiation(this);
-
-        public override bool IsConstructedGenericType => IsGenericType && !IsGenericTypeDefinition;
 
         public override bool ContainsGenericParameters
         {
@@ -3682,8 +3346,6 @@ namespace System
 
         #region Misc
         public sealed override bool HasSameMetadataDefinitionAs(MemberInfo other) => HasSameMetadataDefinitionAsCore<RuntimeType>(other);
-
-        public override bool IsTypeDefinition => RuntimeTypeHandle.IsTypeDefinition(this);
 
         public override Type MakePointerType() => new RuntimeTypeHandle(this).MakePointer();
 
@@ -3830,25 +3492,6 @@ namespace System
             }
 
             throw new ArgumentException(SR.Format(SR.Arg_ObjObjEx, value.GetType(), this));
-        }
-
-        // GetDefaultMembers
-        // This will return a MemberInfo that has been marked with the [DefaultMemberAttribute]
-        public override MemberInfo[] GetDefaultMembers()
-        {
-            // See if we have cached the default member name
-            MemberInfo[] members = null!;
-
-            string? defaultMemberName = GetDefaultMemberName();
-            if (defaultMemberName != null)
-            {
-                members = GetMember(defaultMemberName);
-            }
-
-            if (members == null)
-                members = Array.Empty<MemberInfo>();
-
-            return members;
         }
 
         [DebuggerStepThrough]
@@ -4231,65 +3874,7 @@ namespace System
 
         #endregion
 
-        #region Object Overrides
-
-        public override bool Equals(object? obj)
-        {
-            // ComObjects are identified by the instance of the Type object and not the TypeHandle.
-            return obj == (object)this;
-        }
-
-        public override int GetHashCode() => RuntimeHelpers.GetHashCode(this);
-
         public override string ToString() => GetCachedName(TypeNameKind.ToString)!;
-
-        #endregion
-
-        #region ICloneable
-
-        public object Clone() => this;
-
-        #endregion
-
-        #region ICustomAttributeProvider
-
-        public override object[] GetCustomAttributes(bool inherit)
-        {
-            return CustomAttribute.GetCustomAttributes(this, ObjectType, inherit);
-        }
-
-        public override object[] GetCustomAttributes(Type attributeType, bool inherit)
-        {
-            if (attributeType is null)
-                throw new ArgumentNullException(nameof(attributeType));
-
-            RuntimeType? attributeRuntimeType = attributeType.UnderlyingSystemType as RuntimeType;
-
-            if (attributeRuntimeType == null)
-                throw new ArgumentException(SR.Arg_MustBeType, nameof(attributeType));
-
-            return CustomAttribute.GetCustomAttributes(this, attributeRuntimeType, inherit);
-        }
-
-        public override bool IsDefined(Type attributeType, bool inherit)
-        {
-            if (attributeType is null)
-                throw new ArgumentNullException(nameof(attributeType));
-
-            RuntimeType? attributeRuntimeType = attributeType.UnderlyingSystemType as RuntimeType;
-
-            if (attributeRuntimeType == null)
-                throw new ArgumentException(SR.Arg_MustBeType, nameof(attributeType));
-
-            return CustomAttribute.IsDefined(this, attributeRuntimeType, inherit);
-        }
-
-        public override IList<CustomAttributeData> GetCustomAttributesData()
-        {
-            return CustomAttributeData.GetCustomAttributesInternal(this);
-        }
-
-        #endregion
 
         #region MemberInfo Overrides
 
@@ -4330,22 +3915,7 @@ namespace System
         [MethodImpl(MethodImplOptions.NoInlining)]
         private string? GetCachedName(TypeNameKind kind) => Cache.GetName(kind);
 
-        public override MemberTypes MemberType
-        {
-            get
-            {
-                if (IsPublic || IsNotPublic)
-                    return MemberTypes.TypeInfo;
-                    
-                return MemberTypes.NestedType;
-            }
-        }
-
         public override Type? DeclaringType => Cache.GetEnclosingType();
-
-        public override Type? ReflectedType => DeclaringType;
-
-        public override int MetadataToken => RuntimeTypeHandle.GetToken(this);
 
         #endregion
